[TAN-1106] Move user role code to concern
